requirements-dev.txt
-e .

django==2
django-webpack-loader>=0.5
djangorestframework>=3.7
djangorestframework-gis
django-filter
django-test-plus
django-money
django-reversion
django-reversion-compare
diff-match-patch
bibtexparser>=0.5.5

git+https://github.com/MaxBo/django-publications-bootstrap.git#egg=django-publications-bootstrap>=2.4.1

drf-nested-routers
numpy
matplotlib
plotly
psycopg2
geojson

django-debug-toolbar

django-dynamic-fixture
git+https://github.com/MaxBo/django-fixture-magic.git#egg=django-fixture-magic>=0.1.4

fake-factory
factory_boy
coreapi

mock
pytest
pytest-cov
pytest-django
pytest-splinter
pytest-xdist
codecov
gitpython
selenium